Moved CleanDateToSpan from pipeline.projects.aata to pipeline.nodes.basic.
